allow culling page load stats
Keeping a record of every single page load does not work well
with the accumulated storage pattern.
By accessing /cull-page-stats, one can manually clean up the entries
with low counts, with a threshold configured through env variables.

Of course, this is a bit hacky, as is the entire page load counting.
It's good enough to experiment with and small enough that replacing
it with something better would still be easy.